bidirectional-dijkstra/requirements.txt
numpy==1.24.3  # Numerical computing for grid operations
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
